fix: return element changes without edges
If we modify graph by only changing color of nodes. For example adding

        example:a ui:color "green"

we need to return an element corresponding to example:a

Did a lot of refactoring so the change is hard to spot. The actual
change is:

        const uiEdges = Object.keys(displayPredicate2UiEdges).flatMap((k) => displayPredicate2UiEdges[k]);
        const uiSubjects = uiEdges.map((uiEdge) => uiEdge.from);
        ...
        .concat(uiSubjects)

This let us keep the uiSubjects even if they are not part of any other
edges.

The refactoring is:

* The indirection we get when trying we get when navigating code with
hooks i confusing. Exporting pure functions directly from mapping-code
instead

* Removed some unused code releated to node creations.
* Removed NodeDescription, might be readded

Testing:

* Added turtle_parser test. We now have a single method that maps
turtles to cytoscape elements. This is a great target for testing. The
last test verifies the change mentioned above.
