--- src/hooks/index.ts
@@ -1,3 +1,0 @@
-export * from './cytoscapeHelper';
-export * from './displayControllingPredicate';
-export * from './uiEdge';
