+++ src/index.ts
@@ -1,3 +1,2 @@
 export * from './components';
-export * from './hooks';
 export * from './utils';
