+++ src/mapper/index.ts
@@ -1,3 +1,4 @@
 export * from './httpResult2SparqlResult';
-export * from './useTurtleHelpers';
-export * from './uri2Node';
+export * from './turtle2RdfTriples';
+export * from './rdfTriples2Elements';
+export * from './turtle2Elements';
