--- src/models/nodeDescription.ts
@@ -1,4 +1,0 @@
-export interface NodeDescription {
-	nodeUri: string;
-	description: string;
-}
